+++ GrosirTubesPBO/src/view/component/SupplierForm.java
@@ -86,11 +86,22 @@ public class SupplierForm extends javax.swing.JPanel {
             return;
         }
         
+        int id = Integer.parseInt(tblSupplier.getValueAt(row, 0).toString());
+        
+        // Cek apakah supplier ini terhubung dengan produk
+        if (dao.isSupplierDigunakan(id)) {
+            JOptionPane.showMessageDialog(this, 
+                "Tidak dapat dihapus guna menjaga konsistensi data! \nData ini sudah tercatat dalam laporan.", 
+                "Peringatan", 
+                JOptionPane.WARNING_MESSAGE);
+            return; // STOP, jangan lanjut hapus
+        }
+        
         int confirm = JOptionPane.showConfirmDialog(this, "Hapus data ini?", "Konfirmasi", JOptionPane.YES_NO_OPTION);
         if (confirm == JOptionPane.YES_OPTION) {
-            int id = Integer.parseInt(tblSupplier.getValueAt(row, 0).toString());
             dao.deleteSupplier(id);
             loadData();
+            JOptionPane.showMessageDialog(this, "Data berhasil dihapus!");
         }
     }
